Delete the standalone QA and analysis agents
