+++ docs/agents/analysis.md
@@ -1,6 +1,6 @@
-# Analysis Agent
+# Analysis
 
-The analysis agent enables complex analytical tasks by writing and executing Python code in a sandboxed environment. It solves problems that traditional RAG struggles with:
+The analysis skill enables complex analytical tasks by writing and executing Python code in a sandboxed environment. It solves problems that traditional RAG struggles with:
 
 - **Aggregation**: "How many documents mention security vulnerabilities?"
 - **Computation**: "What's the average revenue across all quarterly reports?"
@@ -9,10 +9,10 @@ The analysis agent enables complex analytical tasks by writing and executing Pyt
 
 ## How It Works
 
-1. The agent receives a question
+1. The skill receives a question
 2. It writes Python code to explore the knowledge base
-3. Code executes in a sandboxed Python interpreter with access to search, LLM, and a virtual filesystem of documents
-4. The agent iterates: run code, examine results, refine approach
+3. Code executes in a sandboxed Python interpreter with access to search and a virtual filesystem of documents
+4. The skill iterates: run code, examine results, refine approach
 5. Final answer is synthesized from the gathered data
 
 ## CLI Usage
@@ -21,11 +21,8 @@ The analysis agent enables complex analytical tasks by writing and executing Pyt
 # Basic usage
 haiku-rag analyze "How many documents are in the database?"
 
-# With document filter (restricts what the agent can access)
+# With document filter (restricts what the skill can access)
 haiku-rag analyze "Summarize the key points" --filter "uri LIKE '%report%'"
-
-# Pre-load specific documents
-haiku-rag analyze "Compare these two reports" --document "Q1 Report" --document "Q2 Report"
 ```
 
 ## Python Usage
@@ -34,23 +31,19 @@ haiku-rag analyze "Compare these two reports" --document "Q1 Report" --document 
 from haiku.rag.client import HaikuRAG
 
 async with HaikuRAG(path_to_db) as client:
-    # Basic question
     result = await client.analyze("How many documents mention 'security'?")
-    print(result.answer)    # The answer
-    print(result.program)   # The final consolidated program
+    print(result.answer)
+    for citation in result.citations:
+        print(citation.uri, citation.title)
 
-    # With filter (agent can only see filtered documents)
+    # With filter (skill can only see filtered documents)
     result = await client.analyze(
         "What is the total revenue?",
         filter="title LIKE '%Financial%'"
     )
-
-    # Pre-load specific documents
-    result = await client.analyze(
-        "Compare the conclusions",
-        documents=["Report A", "Report B"]
-    )
 ```
+
+The executed Python program(s) for each turn are not on `AnalysisResult` itself; they live on `AnalysisState.executions` while the skill runs.
 
 ## Sandbox Capabilities
 
@@ -60,9 +53,8 @@ The agent's code runs in a sandboxed Python interpreter ([pydantic-monty](https:
 
 | Function | Description |
 |----------|-------------|
-| `search(query, limit)` | Hybrid search (vector + full-text) with automatic context expansion. Returns `doc_item_refs` for cross-referencing with `items.jsonl` |
+| `search(query, limit)` | Hybrid search (vector + full-text) with automatic context expansion. Returns `doc_item_refs` and `picture_refs` for cross-referencing with `items.jsonl` |
 | `list_documents()` | List all documents in the knowledge base |
-| `llm(prompt)` | Call an LLM for classification, summarization, or extraction |
 
 ### Document Filesystem
 
@@ -81,13 +73,11 @@ All documents are mounted as a virtual filesystem at `/documents/`. The agent us
 
 Search results include `doc_item_refs` (e.g. `["#/texts/5", "#/tables/0"]`) that match `self_ref` values in `items.jsonl`, enabling navigation from search hits to document structure.
 
-When documents are pre-loaded via the `documents` parameter, they are also injected as a `documents` variable accessible in the sandbox code.
-
 ### Python Features
 
 The interpreter supports a subset of Python: variables, arithmetic, strings, f-strings, lists, dicts, tuples, sets, loops, conditionals, comprehensions, functions, async/await, `filter()`, `getattr()`, try/except, file I/O via `pathlib.Path`, and the `json`, `re`, `math` modules.
 
-Not supported: most imports (only `json`, `re`, `math`, `pathlib` are available), class definitions, generators/yield, match statements, decorators, `with` statements. For pattern matching, the agent can use `import re`, string methods, or the `llm()` function.
+Not supported: most imports (only `json`, `re`, `math`, `pathlib` are available), class definitions, generators/yield, match statements, decorators, `with` statements. For pattern matching, use `import re` or string methods.
 
 ### Security
 
@@ -101,10 +91,10 @@ Code executes in an isolated interpreter with:
 
 ## Context Filter
 
-The `filter` parameter restricts what documents the agent can access. Unlike tool parameters, the filter is applied automatically and cannot be bypassed by the LLM — both the VFS and search results are scoped to the filter:
+The `filter` parameter restricts what documents the skill can access. Unlike tool parameters, the filter is applied automatically and cannot be bypassed by the LLM — both the VFS and search results are scoped to the filter:
 
 ```python
-# Agent can only see documents with "confidential" in the URI
+# Skill can only see documents with "confidential" in the URI
 result = await client.analyze(
     "Summarize all findings",
     filter="uri LIKE '%confidential%'"
