+++ docs/cli.md
@@ -171,21 +171,15 @@ Ask questions about your documents:
 haiku-rag ask "Who is the author of haiku.rag?"
 ```
 
-Ask questions with citations showing source documents:
-```bash
-haiku-rag ask "Who is the author of haiku.rag?" --cite
-```
-
 Filter to specific documents:
 ```bash
 haiku-rag ask "What are the main findings?" --filter "uri LIKE '%paper%'"
 ```
 
-The QA agent searches your documents for relevant information and provides a comprehensive answer. When available, citations use the document title; otherwise they fall back to the URI.
+`ask` runs the [rag skill](skills/index.md) and always renders citations under the answer. When available, citations use the document title; otherwise they fall back to the URI.
 
 Flags:
 
-- `--cite`: Include citations showing which documents were used
 - `--filter` / `-f`: Restrict searches to documents matching the filter (see [Filtering Search Results](python.md#filtering-search-results))
 
 ## Chat
@@ -271,18 +265,11 @@ Filter to specific documents:
 haiku-rag analyze "What is the total revenue?" --filter "title LIKE '%Financial%'"
 ```
 
-Pre-load specific documents for comparison:
-
-```bash
-haiku-rag analyze "Compare the conclusions" --document "Report A" --document "Report B"
-```
-
 Flags:
 
 - `--filter` / `-f`: SQL WHERE clause to restrict document access
-- `--document` / `-d`: Pre-load a document by title or ID (can repeat)
 
-See [Analysis Agent](agents/analysis.md) for details on capabilities and configuration.
+See [Analysis](agents/analysis.md) for details on capabilities and configuration.
 
 ## Create Skill
