+++ docs/configuration/prompts.md
@@ -1,19 +1,16 @@
 # Prompt Customization
 
-Customize the prompts used by haiku.rag's AI agents to better match your domain and use case.
+Customize the prompts used by haiku.rag's skills and research workflow to better match your domain and use case.
 
 ## Configuration
 
 ```yaml
 prompts:
-  # Domain context prepended to all agent prompts
+  # Domain context prepended to skill instructions and research prompts
   domain_preamble: |
     This knowledge base contains technical documentation for the Helios solar panel
     system, including installation manuals, maintenance procedures, and safety guidelines.
     Questions about "the system" or unqualified specs refer to the Helios panel.
-
-  # Full replacement for QA agent prompt (optional)
-  qa: null
 
   # Full replacement for research synthesis prompt (optional)
   synthesis: null
@@ -24,13 +21,13 @@ prompts:
 
 ## Domain Preamble
 
-The `domain_preamble` field provides **domain context** that is prepended to all agent prompts — the main agent, skill subagents, and internal agents (QA, research planning, search, evaluation, and synthesis). Use this to:
+The `domain_preamble` field provides **domain context** prepended to the rag and rag-analysis skill instructions and to the research planner/search/synthesis prompts. Use this to:
 
 - Describe what the knowledge base contains
 - Clarify domain-specific terminology
-- Provide context that helps agents interpret ambiguous queries
+- Provide context that helps the model interpret ambiguous queries
 
-**Important:** `domain_preamble` is for domain context, not behavioral instructions. Descriptions of subject matter, terminology, and content scope belong here. Behavioral guidance (tone, response style, formatting rules) belongs in the agent's system prompt or custom `prompts.qa`.
+**Important:** `domain_preamble` is for domain context, not behavioral instructions. Descriptions of subject matter, terminology, and content scope belong here. Behavioral guidance (tone, response style, formatting rules) lives in the skill's SKILL.md — fork the skill via `haiku-rag create-skill` to customize behavior.
 
 **Example:**
 
@@ -40,34 +37,6 @@ prompts:
     This knowledge base contains product documentation, API references,
     and troubleshooting guides for Acme Corp's cloud platform.
     "Deployment" refers to Acme's managed deployment service, not general CI/CD.
-```
-
-## Custom QA Prompt
-
-Replace the default QA agent prompt entirely by setting `prompts.qa`. The prompt should instruct the agent how to:
-
-1. Use the `search_documents` tool to find relevant content
-2. Interpret search results with scores and metadata
-3. Cite sources using chunk IDs
-4. Handle insufficient information
-
-**Example:**
-
-```yaml
-prompts:
-  qa: |
-    You are a concise technical assistant. Answer questions using only the knowledge base.
-
-    Process:
-    1. Search for relevant documents using the search_documents tool
-    2. Review results ordered by relevance (rank 1 = most relevant)
-    3. Provide a brief, direct answer based on retrieved content
-
-    Guidelines:
-    - Use only information from search results
-    - Include chunk IDs in cited_chunks for sources you use
-    - If information is insufficient, say so clearly
-    - Be concise - avoid unnecessary elaboration
 ```
 
 ## Custom Synthesis Prompt
@@ -130,7 +99,6 @@ from haiku.rag.config.models import PromptsConfig
 config = AppConfig(
     prompts=PromptsConfig(
         domain_preamble="This knowledge base contains Acme Corp product documentation and API references.",
-        qa=None,  # Use default QA prompt
         synthesis=None,  # Use default synthesis prompt
         picture_description="Describe this image for search indexing.",
     )
