+++ haiku_rag_slim/haiku/rag/sandbox/dependencies.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class AnalysisContext:
+    """Mutable context accumulating data during analysis execution."""
+
+    filter: str | None = None
