+++ tests/cassettes/test_client_analyze/TestClientAnalysisIntegration.test_analyze_aggregation.yaml
