+++ tests/sandbox/test_models.py
@@ -1,0 +1,8 @@
+from haiku.rag.sandbox import AnalysisResult
+
+
+class TestAnalysisResult:
+    def test_create_result(self):
+        result = AnalysisResult(answer="The answer is 42")
+        assert result.answer == "The answer is 42"
+        assert result.citations == []
