+++ tests/sandbox/test_sandbox_multimodal.py
@@ -4,10 +4,9 @@ and the set of external functions exposed to the interpreter.
 
 import pytest
 
-from haiku.rag.agents.analysis.dependencies import AnalysisContext
-from haiku.rag.agents.analysis.sandbox import Sandbox
 from haiku.rag.client import HaikuRAG
 from haiku.rag.config.models import AppConfig
+from haiku.rag.sandbox import AnalysisContext, Sandbox
 from haiku.rag.store.models.chunk import SearchResult
